refactor traits
